reorganize extension-configuration
