+++ src/components/ucp-tabs/common/importing/import-strategies/full-strategy.ts
@@ -3,7 +3,7 @@ import { ConfigFile, Extension } from '../../../../../config/ucp/common';
 import { deserializeLoadOrder } from '../../../../../config/ucp/config-files/load-order';
 import { ExtensionsState } from '../../../../../function/extensions/extensions-state';
 import { MessageType } from '../../../../../localization/localization';
-import { buildExtensionConfigurationDB } from '../../../extension-manager/extension-configuration';
+import { buildExtensionConfigurationDB } from '../../../extension-manager/extension-configuration/extension-configuration';
 import {
   GenericFailure,
   LOGGER,
