+++ src/components/ucp-tabs/extension-manager/extension-configuration/tests/extensions-state.json
