+++ src/components/ucp-tabs/extension-manager/extension-elements/customize-extension-button-callback.ts
@@ -20,7 +20,7 @@ import { CONFIG_DIRTY_STATE_ATOM } from '../../common/buttons/config-serialized-
 import { EXTENSION_EDITOR_STATE_ATOM } from '../../common/extension-editor/extension-editor-state';
 import { constructUserConfigObjects } from '../../common/importing/import-button-callback';
 import saveConfig from '../../common/save-config';
-import { buildExtensionConfigurationDB } from '../extension-configuration';
+import { buildExtensionConfigurationDB } from '../extension-configuration/extension-configuration';
 import { addExtensionToExplicityActivatedExtensions } from '../extensions-state-manipulation';
 
 // eslint-disable-next-line import/prefer-default-export
