+++ src/function/game-folder/modifications/activate-first-time-use-extensions.ts
@@ -1,4 +1,4 @@
-import { buildExtensionConfigurationDB } from '../../../components/ucp-tabs/extension-manager/extension-configuration';
+import { buildExtensionConfigurationDB } from '../../../components/ucp-tabs/extension-manager/extension-configuration/extension-configuration';
 import { addExtensionToExplicityActivatedExtensions } from '../../../components/ucp-tabs/extension-manager/extensions-state-manipulation';
 import { Extension } from '../../../config/ucp/common';
 import Logger from '../../../util/scripts/logging';
